+++ cordapp/src/main/kotlin/net/corda/plugins/CordappPlugin.kt
@@ -1,10 +1,7 @@
 package net.corda.plugins
 
-import net.corda.plugins.SigningOptions.Companion.DEFAULT_KEYSTORE
-import net.corda.plugins.SigningOptions.Companion.DEFAULT_KEYSTORE_EXTENSION
-import net.corda.plugins.SigningOptions.Companion.DEFAULT_KEYSTORE_FILE
+import net.corda.plugins.SignJar.Companion.sign
 import net.corda.plugins.Utils.Companion.compareVersions
-import net.corda.plugins.Utils.Companion.createTempFileFromResource
 import org.gradle.api.GradleException
 import org.gradle.api.InvalidUserDataException
 import org.gradle.api.Plugin
@@ -13,7 +10,6 @@ import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.Dependency
 import org.gradle.jvm.tasks.Jar
 import java.io.File
-import java.nio.file.Paths
 
 /**
  * The Cordapp plugin will turn a project into a cordapp project which builds cordapp JARs with the correct format
@@ -42,8 +38,7 @@ class CordappPlugin : Plugin<Project> {
         Utils.createCompileConfiguration("cordaCompile", project)
         Utils.createRuntimeConfiguration("cordaRuntime", project)
 
-        cordapp = project.extensions.create("cordapp", CordappExtension::class.java)
-        cordapp.setProject(project)
+        cordapp = project.extensions.create("cordapp", CordappExtension::class.java, project.objects)
         configureCordappJar(project)
     }
 
@@ -57,9 +52,9 @@ class CordappPlugin : Plugin<Project> {
         jarTask.doFirst {
             val (targetPlatformVersion, minimumPlatformVersion) = checkVersionInfo()
             val attributes = jarTask.manifest.attributes
-            attributes["Name"] = cordapp.info?.name ?: "${project.group}.${jarTask.baseName}"
-            attributes["Implementation-Version"] = cordapp.info?.version ?: project.version
-            attributes["Implementation-Vendor"] = cordapp.info?.vendor ?: UNKNOWN
+            attributes["Name"] = cordapp.info.name ?: "${project.group}.${jarTask.baseName}"
+            attributes["Implementation-Version"] = cordapp.info.version ?: project.version
+            attributes["Implementation-Vendor"] = cordapp.info.vendor ?: UNKNOWN
             attributes["Target-Platform-Version"] = targetPlatformVersion
             attributes["Min-Platform-Version"] = minimumPlatformVersion
             if (attributes["Implementation-Vendor"] == UNKNOWN) {
@@ -69,7 +64,7 @@ class CordappPlugin : Plugin<Project> {
                 attributes["Sealed"] = "true"
             }
         }.doLast {
-            sign(project, cordapp.signing)
+            sign(project, cordapp.signing, it.outputs.files.singleFile)
         }
         task.doLast {
             jarTask.from(getDirectNonCordaDependencies(project).map {
@@ -85,37 +80,6 @@ class CordappPlugin : Plugin<Project> {
             }
         }
         jarTask.dependsOn(task)
-    }
-
-    private fun sign(project: Project, signing: Signing) {
-        if (!signing.enabled) {
-            project.logger.info("CorDapp JAR signing is disabled, the CorDapp's contracts will not use signature constraints.")
-            return
-        }
-        val options = signing.options.toSignJarOptionsMap()
-        if (signing.options.hasDefaultOptions()) {
-            project.logger.info("CorDapp JAR signing with the default Corda development key, suitable for Corda running in development mode only.")
-            val keyStorePath = createTempFileFromResource(DEFAULT_KEYSTORE, DEFAULT_KEYSTORE_FILE, DEFAULT_KEYSTORE_EXTENSION)
-            options["keystore"] = keyStorePath.toString()
-        }
-
-        val path = project.tasks.getByName("jar").outputs.files.singleFile.toPath()
-        options["jar"] = path.toString()
-
-        try {
-            project.ant.invokeMethod("signjar", options)
-        } catch (e: Exception) {
-            // Not adding error message as it's always meaningless, logs with --INFO level contain more insights
-            throw InvalidUserDataException("Exception while signing ${path.fileName}, " +
-                    "ensure the 'cordapp.signing.options' entry contains correct keyStore configuration, " +
-                    "or disable signing by 'cordapp.signing.enabled false'. " +
-                    if (project.logger.isInfoEnabled || project.logger.isDebugEnabled) "Search for 'ant:signjar' in log output."
-                    else "Run with --info or --debug option and search for 'ant:signjar' in log output. ", e)
-        } finally {
-            if (signing.options.hasDefaultOptions()) {
-                Paths.get(options["keystore"]).toFile().delete()
-            }
-        }
     }
 
     private fun getDirectNonCordaDependencies(project: Project): Set<File> {
@@ -154,8 +118,9 @@ class CordappPlugin : Plugin<Project> {
 
     private fun checkVersionInfo(): Pair<Int, Int> {
         // If the minimum platform version is not set, default to 1.
-        val minimumPlatformVersion: Int = cordapp.info?.minimumPlatformVersion ?: 1
-        val targetPlatformVersion = cordapp.info?.targetPlatformVersion ?: throw InvalidUserDataException("Target version was not set and could not be determined from the project's Corda dependency. Please specify the target version of your CorDapp.")
+        val minimumPlatformVersion: Int = cordapp.info.minimumPlatformVersion ?: 1
+        val targetPlatformVersion = cordapp.info.targetPlatformVersion
+                ?: throw InvalidUserDataException("Target version was not set and could not be determined from the project's Corda dependency. Please specify the target version of your CorDapp.")
         if (targetPlatformVersion < 1) {
             throw InvalidUserDataException("Target version must not be smaller than 1.")
         }