+++ cordapp/src/main/kotlin/net/corda/plugins/Info.kt
@@ -1,3 +1,16 @@
 package net.corda.plugins
 
-open class Info(var name: String? = null, var version: String? = null, var vendor: String? = null, var targetPlatformVersion: Int? = null, var minimumPlatformVersion: Int? = null)
+import org.gradle.api.tasks.Input
+
+open class Info {
+    @get:Input
+    var name: String? = null
+    @get:Input
+    var version: String? = null
+    @get:Input
+    var vendor: String? = null
+    @get:Input
+    var targetPlatformVersion: Int? = null
+    @get:Input
+    var minimumPlatformVersion: Int? = null
+}