+++ changelog.md
@@ -7,6 +7,7 @@
 ### Version 4.0.33
 
 * `cordformation`: Add ability to specify a flow override as part of cordformation via `flowOverride`
+* `cordapp`: New `SignJar` task allows to sign arbitrary JAR files reusing `signing` configuration of the `Cordapp` plugin
 
 ### Version 4.0.32